Fixing dtype use
Former-commit-id: c6b51c1ccd2afc39e48bb96dd2ac9095b51f0eda

+++ SHE_GST_PrepareConfigs/python/SHE_GST_PrepareConfigs/write_configs.py
@@ -140,8 +140,7 @@ def write_configs_from_plan( plan_filename,
         model_seeds = np.linspace(start=mseed_min,
                                   stop=mseed_max,
                                   num=(mseed_max-mseed_min)//mseed_step+1,
-                                  endpoint=True,
-                                  dtype=int)
+                                  endpoint=True)
         
         nseed_min = noise_seed_mins[row_index]
         nseed_max = noise_seed_maxes[row_index]
